+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2025-09-08
+
+### Added
+
+- Added Event pack
+
+### Changed
+
+- Update to MicroEJ SDK gradle plugin 1.3.1
+- MicroEJ Architecture updated to version 8.4.0
+- MicroEJ Core test suite updated to 3.5.0
+- MicroEJ UI Architecture updated to version 14.4.2
+- Update to OSAL 1.0.3 and Async-worker 0.6.0
+- Implement event queue on top of OSAL_queue instead of mqueue
+
+### Fixed
+
+- Fixed a timeout in osal posix
+- Fixed run.sh script for application.out deployment
+
 ## [3.0.2] - 2025-03-28
 
 ### Fixed